external_agents/fred/requirements.txt
# FRED Agent Dependencies
# Install: pip install -r requirements.txt

# Core dependencies
fredapi==0.5.2              # Official FRED API wrapper
pandas==2.2.0               # Data manipulation
numpy==1.26.3               # Numerical operations
python-dotenv==1.0.0        # Environment variable management

# ADK and A2A
google-adk>=0.1.0           # Agent Development Kit

# Utilities
requests==2.31.0            # HTTP requests (backup for fredapi)
cachetools==5.3.2           # Response caching
python-dateutil==2.8.2      # Date handling

# Optional: Development/Testing
pytest==7.4.4               # Testing framework
pytest-asyncio==0.23.3      # Async testing support
